+++ CHANGELOG.md
@@ -20,6 +20,8 @@ any parts of the framework not mentioned in the documentation should generally b
 * Don't swallow `filter[]` params when there are several
 * Fix DeprecationWarning regarding collections.abc import in Python 3.7
 * Allow OPTIONS request to be used on RelationshipView.
+* Remove non-JSONAPI methods (PUT and TRACE) from ModelViewSet and RelationshipView. This is a **BREAKING CHANGE** if
+  your clients are incorrectly using PUT instead of PATCH.
 
 ## [2.7.0] - 2019-01-14